rename: vending machine components move
