+++ src/components/VendingMachine/VMInputBox.jsx
