+++ src/components/VendingMachine/VMItem.jsx
