+++ src/components/VendingMachine/VMLogs.jsx
